Updated based on review feedback
