--- scripts/helpers/build_tournament_match_file.py
@@ -1,4 +1,0 @@
-import helper_utils
-
-
-helper_utils.read_all_matches()
